Move output (pdf/html/image) related code to utilities.output.
